refactor layout so that scroll down button does not over lap.
